exceptions + docker + docker compose

+++ core/apps/posts/services/posts.py
@@ -3,6 +3,7 @@ from abc import (
     abstractmethod,
 )
 
+from core.apps.posts.exceptions import ForbiddenException, PostNotFoundException
 from core.apps.posts.models import Post as PostModel
 from core.apps.posts.entities import Post as PostEntity
 
@@ -29,7 +30,11 @@ class BasePostService(ABC):
 
 class PostService(BasePostService):
     def get(self, post_id: int) -> PostEntity:
-        post_model = PostModel.objects.get(id=post_id)
+        try:
+            post_model = PostModel.objects.get(id=post_id)
+        except PostModel.DoesNotExist:
+            raise PostNotFoundException()
+
         return post_model.to_entity()
 
     def create(self, title: str, content: str, user_id: int) -> PostEntity:
@@ -48,13 +53,30 @@ class PostService(BasePostService):
         post_models = PostModel.objects.order_by("-created_at")
         return [post_model.to_entity() for post_model in post_models]
 
-    def update(self, post_id: int, title: str, content: str) -> PostEntity:
-        post_model = PostModel.objects.get(id=post_id)
+    def update(
+        self, post_id: int, user_id: int, title: str, content: str
+    ) -> PostEntity:
+        try:
+            post_model = PostModel.objects.get(id=post_id)
+        except PostModel.DoesNotExist:
+            raise PostNotFoundException()
+
+        if user_id != post_model.user_id:
+            raise ForbiddenException()
+
         post_model.title = title
         post_model.content = content
         post_model.save()
+
         return post_model.to_entity()
 
-    def delete(self, post_id: int) -> None:
-        post_model = PostModel.objects.get(id=post_id)
+    def delete(self, post_id: int, user_id: int) -> None:
+        try:
+            post_model = PostModel.objects.get(id=post_id)
+        except PostModel.DoesNotExist:
+            raise PostNotFoundException()
+
+        if user_id != post_model.user_id:
+            raise ForbiddenException()
+
         post_model.delete()
